+++ lib/src/request/keychain_derive_address.dart
@@ -1,0 +1,16 @@
+/// SPDX-License-Identifier: AGPL-3.0-or-later
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'keychain_derive_address.freezed.dart';
+part 'keychain_derive_address.g.dart';
+
+@freezed
+class KeychainDeriveAddressResult with _$KeychainDeriveAddressResult {
+  const factory KeychainDeriveAddressResult({
+    required String address,
+  }) = _KeychainDeriveAddressResult;
+  const KeychainDeriveAddressResult._();
+
+  factory KeychainDeriveAddressResult.fromJson(Map<String, dynamic> json) =>
+      _$KeychainDeriveAddressResultFromJson(json);
+}