+++ lib/src/request/keychain_derive_address.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'keychain_derive_address.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_KeychainDeriveAddressResult _$$_KeychainDeriveAddressResultFromJson(
+        Map<String, dynamic> json) =>
+    _$_KeychainDeriveAddressResult(
+      address: json['address'] as String,
+    );
+
+Map<String, dynamic> _$$_KeychainDeriveAddressResultToJson(
+        _$_KeychainDeriveAddressResult instance) =>
+    <String, dynamic>{
+      'address': instance.address,
+    };