+++ lib/src/request/keychain_derive_keypair.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'keychain_derive_keypair.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_KeychainDeriveKeypairResult _$$_KeychainDeriveKeypairResultFromJson(
+        Map<String, dynamic> json) =>
+    _$_KeychainDeriveKeypairResult(
+      publicKey: json['publicKey'] as String,
+    );
+
+Map<String, dynamic> _$$_KeychainDeriveKeypairResultToJson(
+        _$_KeychainDeriveKeypairResult instance) =>
+    <String, dynamic>{
+      'publicKey': instance.publicKey,
+    };